+++ crack-core/src/commands/music/queue.rs
@@ -2,11 +2,9 @@ use self::serenity::{builder::CreateEmbed, futures::StreamExt};
 use crate::errors::CrackedError;
 use crate::{
     handlers::track_end::ModifyQueueHandler,
+    interface::build_nav_btns,
     messaging::messages::QUEUE_EXPIRED,
-    utils::{
-        build_nav_btns, calculate_num_pages, create_queue_embed, forget_queue_message,
-        get_interaction,
-    },
+    utils::{calculate_num_pages, create_queue_embed, forget_queue_message, get_interaction},
     Context, Error,
 };
 use ::serenity::builder::{
