+++ crack-core/src/handlers/track_end.rs
@@ -9,9 +9,10 @@ use crate::{
     commands::{enqueue_track_pgwrite, forget_skip_votes, send_now_playing, MyAuxMetadata},
     db::PlayLog,
     errors::{verify, CrackedError},
+    interface::build_nav_btns,
     messaging::messages::SPOTIFY_AUTH_FAILED,
     sources::spotify::{Spotify, SPOTIFY},
-    utils::{build_nav_btns, calculate_num_pages, create_queue_embed, forget_queue_message},
+    utils::{calculate_num_pages, create_queue_embed, forget_queue_message},
     Data,
 };
